requirements.txt
psycopg2-binary
pytz
requests
